refactor: move test helpers to data
