+++ src/app/resume-page/data/__tests__/make-json-resume-project.ts
@@ -1,0 +1,13 @@
+import resume from '@/data/resume.json'
+import { JsonResumeProject } from '../../projects-section/json-resume-projects'
+
+const sampleJsonResumeProject = resume.projects[0]
+
+export function makeJsonResumeProject(
+  overrides: Partial<JsonResumeProject> = {},
+): JsonResumeProject {
+  return {
+    ...sampleJsonResumeProject,
+    ...overrides,
+  } as JsonResumeProject
+}
