+++ src/app/resume-page/experience-section/experience/experience-to-contents.spec.ts
@@ -3,7 +3,7 @@ import { experienceToContents } from './experience-to-contents'
 import { makeExperience } from '../../data/__tests__/make-experience'
 import { ExperienceHighlightsComponent } from './experience-highlights/experience-highlights.component'
 import { Project } from '../../data/project'
-import { makeProject } from '../../projects-section/__tests__/make-project'
+import { makeProject } from '../../data/__tests__/make-project'
 import { ExperienceTechComponent } from './experience-tech/experience-tech.component'
 import { ChippedContent } from '../../chipped-content/chipped-content'
