+++ src/app/resume-page/projects-section/projects-section.component.spec.ts
@@ -6,7 +6,7 @@ import { SectionTitleComponent } from '../section-title/section-title.component'
 import { GET_PROJECT_ITEMS, GetProjectItems } from './get-project-items'
 import { ProjectItemComponent } from './project-item/project-item.component'
 import { componentTestSetup } from '@/test/helpers/component-test-setup'
-import { makeProject } from './__tests__/make-project'
+import { makeProject } from '../data/__tests__/make-project'
 import { CardGridComponent } from '../card-grid/card-grid.component'
 import { By } from '@angular/platform-browser'
